requirements.txt
# Web / frontend
streamlit>=1.28.0
pydantic>=2.0.0
python-multipart>=0.0.10

# LlamaIndex core + LLMs + embeddings
llama-index-core>=0.13.0
llama-index-llms-llama-cpp>=0.1.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-vector-stores-chroma>=0.1.0
llama-index-vector-stores-faiss>=0.1.0   # FAISS support

# Transformers / embeddings
transformers>=4.44.2
sentence-transformers>=3.1.1
torch>=2.0.0

# Vector DB
chromadb>=0.4.0
faiss-cpu>=1.7.4  # CPU FAISS; use faiss-gpu if you move to CUDA

# PDF / DOCX handling
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.0.0

# Packaging
pyinstaller>=5.0.0